Migrated sources to use Swift Package Manager
