+++ Application/Sources/Dynamic Price/GoogleAdManager.swift
