+++ Application/Sources/Rendering/NimbusAdManager.swift
